# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoCamisetas API",
        "description": "API REST para gestionar camisetas, clientes, tallas y calcular precio final segun el cliente consultado.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/tallas": {
            "get": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Listar tallas",
                "operationId": "getTallas",
                "responses": {
                    "200": {
                        "description": "Listado exitoso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Crear talla",
                "operationId": "createTalla",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "M"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Talla creada"
                    },
                    "422": {
                        "description": "Errores de validacion"
                    }
                }
            }
        },
        "/tallas/{id}": {
            "get": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Obtener talla por ID",
                "operationId": "getTalla",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Talla encontrada"
                    },
                    "404": {
                        "description": "Talla no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Actualizar talla",
                "operationId": "updateTalla",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Talla actualizada"
                    },
                    "404": {
                        "description": "Talla no encontrada"
                    },
                    "422": {
                        "description": "Errores de validacion"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Eliminar talla",
                "operationId": "deleteTalla",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Talla eliminada"
                    },
                    "404": {
                        "description": "Talla no encontrada"
                    },
                    "409": {
                        "description": "Talla asociada a camisetas"
                    }
                }
            }
        },
        "/camisetas/{id}/tallas": {
            "get": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Listar tallas asociadas a una camiseta",
                "operationId": "getTallasByCamiseta",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso"
                    },
                    "404": {
                        "description": "Camiseta no encontrada"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Asociar talla a camiseta",
                "operationId": "attachTallaToCamiseta",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "talla_id"
                                ],
                                "properties": {
                                    "talla_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Talla asociada"
                    },
                    "404": {
                        "description": "Camiseta no encontrada"
                    },
                    "409": {
                        "description": "Talla ya asociada"
                    },
                    "422": {
                        "description": "Errores de validacion"
                    }
                }
            }
        },
        "/camisetas/{id}/tallas/{tallaId}": {
            "delete": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Desasociar talla de camiseta",
                "operationId": "detachTallaFromCamiseta",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "tallaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Talla desasociada"
                    },
                    "404": {
                        "description": "Camiseta, talla o asociacion no encontrada"
                    }
                }
            }
        }
    },
...
}